+++ CHANGELOG.md
@@ -50,7 +50,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
+[Unreleased]: https://github.com/japaric/stack-sizes/compare/v0.3.0...HEAD
 [v0.3.0]: https://github.com/japaric/stack-sizes/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/japaric/stack-sizes/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/japaric/stack-sizes/compare/v0.1.0...v0.1.1
-[Unreleased]: https://github.com/japaric/stack-sizes/compare/v0.1.0...HEAD