+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.3.1] - 2018-03-10
+
+### Added
+
+- `Function` now has a `size` method to get the size of the subroutine in bytes.
+
 ## [v0.3.0] - 2019-03-03
 
 ### Changed
@@ -50,7 +56,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
-[Unreleased]: https://github.com/japaric/stack-sizes/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/japaric/stack-sizes/compare/v0.3.1...HEAD
+[v0.3.1]: https://github.com/japaric/stack-sizes/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/japaric/stack-sizes/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/japaric/stack-sizes/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/japaric/stack-sizes/compare/v0.1.0...v0.1.1